Scene Structure Implementation
Player's will be sent to a Title Screen and await the full loading of the Main Menu allowing assets to be loaded and UI Elements to be initialized behind the curtains.

After it's done they can press SPACE and transition to the Main Menu scene and discard the Title Screen.

This allows for complex menu compositions but also scales well for simple menus.